refactor: Break up components, add subfolders
