+++ editor.planx.uk/src/pages/FlowEditor/components/Settings/ServiceSettings/FlowElements/components/TextInput.tsx
@@ -1,0 +1,69 @@
+import Box from "@mui/material/Box";
+// eslint-disable-next-line no-restricted-imports
+import { SwitchProps } from "@mui/material/Switch";
+import React from "react";
+import RichTextInput from "ui/editor/RichTextInput/RichTextInput";
+import SettingsDescription from "ui/editor/SettingsDescription";
+import Input, { Props as InputProps } from "ui/shared/Input/Input";
+import InputRow from "ui/shared/InputRow";
+import InputRowItem from "ui/shared/InputRowItem";
+import { Switch } from "ui/shared/Switch";
+
+export const TextInput: React.FC<{
+  title: string;
+  richText?: boolean;
+  description?: string;
+  switchProps?: SwitchProps;
+  headingInputProps?: InputProps;
+  contentInputProps?: InputProps;
+}> = ({
+  title,
+  richText = false,
+  description,
+  switchProps,
+  headingInputProps,
+  contentInputProps,
+}) => {
+  return (
+    <Box width="100%">
+      <Box mb={0.5} display="flex" alignItems="center">
+        <Switch
+          label={title}
+          name={switchProps?.name}
+          variant="editorPage"
+          onChange={switchProps?.onChange}
+          checked={switchProps?.checked}
+        />
+      </Box>
+      <Box mb={1}>
+        {description && (
+          <SettingsDescription>{description}</SettingsDescription>
+        )}
+      </Box>
+      <InputRow>
+        <InputRowItem>
+          <Input placeholder="Heading" format="bold" {...headingInputProps} />
+        </InputRowItem>
+      </InputRow>
+      <InputRow>
+        <InputRowItem>
+          {richText ? (
+            <RichTextInput
+              placeholder="Text"
+              multiline
+              rows={6}
+              {...contentInputProps}
+            />
+          ) : (
+            <Input
+              placeholder="Text"
+              multiline
+              rows={6}
+              {...contentInputProps}
+            />
+          )}
+        </InputRowItem>
+      </InputRow>
+    </Box>
+  );
+};
