+++ editor.planx.uk/src/pages/FlowEditor/components/Settings/ServiceSettings/FlowStatus/tests/testUtils.tsx
@@ -5,7 +5,7 @@ import { HTML5Backend } from "react-dnd-html5-backend";
 import { setup } from "testUtils";
 import { vi } from "vitest";
 
-import ServiceSettings from "..";
+import ServiceSettings from "../../ServiceSettings";
 
 export default async function setupServiceSettingsScreen() {
   const { user } = setup(
