+++ editor.planx.uk/src/pages/FlowEditor/components/Settings/ServiceSettings/ServiceSettings.tsx
@@ -1,8 +1,8 @@
 import Container from "@mui/material/Container";
 import React from "react";
 
+import { FooterLinksAndLegalDisclaimer } from "./FlowElements/FooterLinksAndLegalDisclaimer";
 import FlowStatus from "./FlowStatus";
-import { FooterLinksAndLegalDisclaimer } from "./FooterLinksAndLegalDisclaimer";
 
 const ServiceSettings: React.FC = () => (
   <Container
